pyproject.toml
[tool.poetry]
name = "airflow-project"
version = "0.1.0"
description = ""
authors = ["ToudertiHiba <hibatouderti@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
bs4 = "^0.0.1"
pandas = "2.1.1"
pytest = "^7.4.3"
s3fs = "^2023.10.0"
streamlit = "^1.28.0"
matplotlib = "^3.8.1"
wordcloud = "^1.9.2"
streamlit-plotly-events = "^0.0.6"
apache-airflow = "^2.7.2"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
requests-mock = "^1.11.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
